chg: [Cluster] add(cv: ClusterValue)
