+++ .github/workflows/pytest.yml
@@ -33,7 +33,7 @@ jobs:
 
     - name: Test with nosetests
       run: |
-        poetry run pytest --cov=pymispgalaxies  tests/tests.py
+        poetry run pytest --cov=pymispgalaxies  tests/test*.py
         poetry run mypy .
 
     - name: Upload coverage to Codecov
